TestServerRAG/requirements.txt
fastapi
langserve
langchain-openai
langchain-core
langchain-pinecone
pydantic
uvicorn
sse_starlette
python-dotenv